Refactor: split client main.go into focused files
Split monolithic client main.go into separate files by responsibility:
- net.go: readLoop and writeLoop (TCP read/write goroutines)
- ui.go: clearScreen, moveCursorToBottom, printMessage, promptLogin
- config.go: serverAddr const
- main.go: entry point only, connects and wires everything together
