+++ cmd/client/config.go
@@ -1,0 +1,3 @@
+package main
+
+const serverAddr = "localhost:3000"
